+++ algorithms/follow_ball.py
@@ -1,5 +1,5 @@
 import logging
-import constants
+import core.constants
 import algorithms.heading_hold as hh # hh to match other styles in our code base
