+++ core/logging.py
@@ -1,7 +1,7 @@
 import struct
-import constants
+import core.constants
 import time
-from drivers.rovecomm import RoveCommPacket
+from core.rovecomm import RoveCommPacket
 
 class LogWriter:
     def __init__(self,filename):
